refactor: seperate test class
